refactoring: all pages are located into individual folders and all components which related to page are located into the same folder
