+++ src/screens/ColorDetails/ColorDetailsPage/ColorDetailsPage.module.scss
