+++ src/screens/Colors/ColorCard/ColorCard.jsx
