+++ src/screens/Colors/ColorCard/ColorCard.module.scss
