+++ src/screens/Colors/ColorsPage/ColorsPage.module.scss
