+++ src/screens/Login/Auth-form/Auth-form.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import styles from './Auth-form.module.scss';
 import { Formik } from 'formik';
-import StringButton from '../String-button/String-button';
-import PrimaryButton from '../Primary-button/Primary-button';
+import StringButton from '../../../components/String-button/String-button';
+import PrimaryButton from '../../../components/Primary-button/Primary-button';
 import FormInput from './FormInput';
 
 export default class AuthForm extends React.Component {
