+++ src/screens/Login/Login-form/Login-form.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import styles from './Login-form.module.scss';
 import AuthForm from '../Auth-form/Auth-form';
-import StringButton from '../String-button/String-button';
-import Popup from '../Popup/Popup';
+import StringButton from '../../../components/String-button/String-button';
+import Popup from '../../../components/Popup/Popup';
 
 export default class LoginForm extends React.Component {
