+++ src/screens/Login/LoginPage/LoginPage.jsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from 'react';
-import LoginForm from '../../components/Login-form/Login-form';
-import styles from './Login-page.module.scss';
+import LoginForm from '../Login-form/Login-form';
+import styles from './LoginPage.module.scss';
 import { useNavigate, useLocation } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
-import { createUser, authUser, resetUserWasNotFoundField } from '../../store/user/actions';
-import Popup from '../../components/Popup/Popup';
+import { createUser, authUser, resetUserWasNotFoundField } from '../../../store/user/actions';
+import Popup from '../../../components/Popup/Popup';
 
 const LoginPage = () => {
   const [formTitle] = useState('People Collection');
