+++ src/screens/People/People-page/People-page.module.scss
